A lot of work done offline
Worked offline here and there over the holidays. Changed a few styles, did some work on a static example for extra info on a content card. Changed the methods of a few animations. Now will start work on database connection for an example, aswell as fixup the actual display of data